fix: fix nodes_created counting not as expected

+++ search/models/graph.py
@@ -14,4 +14,4 @@ class Node:
     id: int
     x: float
     y: float
-    neighbors: list[tuple[int, float]] = field(default_factory=list)
+    neighbors: list[tuple[int, int]] = field(default_factory=list)
